+++ src/index.ts
@@ -8,4 +8,5 @@ export * from "./components/Block.js";
 export * from "./components/Breadcrumbs.js";
 export * from "./components/ContentWrapper.js";
 export * from "./components/Header.js";
+export * from "./components/HumanDateTime.js";
 export * from "./components/Notice.js";